src/requirements.txt
flet[all]
flet-rive
flet-lottie
torch
torchvision
torchaudio
numpy
scipy
pandas
dask
pandas-profiling
seaborn
matplotlib
librosa
opencv-python
moviepy
mediapipe
tqdm
kymatio
pillow
llama-cpp-python
bs4
transformers
sentence-transformers
ffmpeg[all]
datasets